Add files via upload
Day 1 goal: "Your goal is to create a function that removes the first and last letters of a string. Strings with two characters or less are considered invalid. You can choose to have your function return null or simply ignore."

I wrote a simple c++ program to do this with the user inputting a string and then returning the string back without the first and last character.